Move ScopedHook to its own file
