+++ projects/rccl/test/host/MICROTEST_README.md
@@ -116,8 +116,11 @@ test:
    target's `test/test_categories_micro*.yaml` so CTest runs it.
 3. **Write the `TEST` / fixture.** Use a fixture whose `TearDown()` calls
    `ResetP2pFakes()` so hooks do not leak between tests. Install per-test
-   behaviour by overwriting a `std::function` hook (see the `ScopedHook`
-   helper in `p2p-test.cc`) rather than editing a fake's default.
+   behaviour by overwriting a `std::function` hook rather than editing a fake's
+   default. Prefer the `ScopedHook` helper in `ScopedHook.h` (see
+   [Installing per-test behaviour with `ScopedHook`](#installing-per-test-behaviour-with-scopedhook)),
+   which installs the hook, counts calls, and restores the previous behaviour
+   automatically on scope exit.
 4. **Only exercise faked seams.** Every external symbol the `#include`d `.cc`
    reaches must be satisfied by `fakes/`: a missing symbol surfaces as a
    link error, a wrongly
@@ -195,6 +198,82 @@ void ResetRmaFakes() {
 }
 ```
 
+### Installing per-test behaviour with `ScopedHook`
+
+Once a seam is a `std::function` hook, install per-test behaviour with the
+`ScopedHook` RAII helper (`test/host/ScopedHook.h`) instead of assigning the
+global directly and remembering to reset it. `ScopedHook` does three things:
+
+1. installs the test's behaviour on construction,
+2. counts calls automatically via its `.calls` member, and
+3. restores the previous behaviour in its destructor — so a hook can't leak
+   into the next test even if you forget the fixture's `TearDown` reset.
+
+Class template argument deduction picks up the signature from the hook
+variable, so call sites don't spell it out:
+
+```cpp
+#include "ScopedHook.h"
+
+TEST_F(P2pMicrotest, IpcRegisterBuffer_UsesBaseAddr) {
+    ScopedHook memGet(g_hipMemGetAddressRange,
+        [&](hipDeviceptr_t* pb, std::size_t* ps, hipDeviceptr_t) {
+            if (pb) *pb = /* canned base addr */;
+            if (ps) *ps = /* canned size */;
+            return hipSuccess;
+        });
+
+    // ... call the unit under test ...
+
+    EXPECT_EQ(memGet.calls, 1);
+    // memGet's destructor restores g_hipMemGetAddressRange here.
+}
+```
+
+`ScopedHook` is intentionally non-copyable and non-movable (the installed
+lambda captures `this` to bump the counter), so declare it as a local; C++17
+guaranteed copy elision lets helper factories still return it by value.
+
+### Naming a reusable seam behaviour with a lambda factory
+
+When the *same* seam behaviour is needed by many tests — e.g. "make this param
+return non-zero so control reaches branch X" — don't copy-paste the lambda into
+every test. Wrap it in a small **lambda-factory helper**: a function that
+*returns a lambda* whose signature matches the seam. This keeps the "what value
+must this seam return, and why" knowledge in one named, commented place.
+
+```cpp
+// Returns a hook lambda that reports the buffer as legacy-IPC-capable, so the
+// `else if (legacyIpcCap)` arm of ipcRegisterBuffer is reachable.
+auto ForceLegacyIpcCapable()
+{
+    return [](void* data, hipPointer_attribute attribute,
+              hipDeviceptr_t) -> hipError_t {
+        if (data && attribute == HIP_POINTER_ATTRIBUTE_IS_LEGACY_HIP_IPC_CAPABLE)
+            *static_cast<int*>(data) = 1;
+        return hipSuccess;
+    };
+}
+```
+
+Return a **plain lambda, not a pre-wrapped `ScopedHook`**, so call sites stay
+free to compose the recipe into whichever installation form they need:
+
+```cpp
+// As a ScopedHook local (CTAD deduces the signature from the hook variable):
+ScopedHook pointerAttr(g_hipPointerGetAttribute, ForceLegacyIpcCapable());
+
+// Or into an optional/emplace form owned by a shared fixture:
+pointerAttr.emplace(g_hipPointerGetAttribute, ForceLegacyIpcCapable());
+```
+
+Name the factory after the *state it forces* (`ForceLegacyIpcCapable`,
+`ForceLegacyCudaRegister`), not after the seam it drives — the point is that a
+test reads as "force this precondition, then call the unit under test". Where a
+single branch depends on HIP version (only one arm is live per build, but the
+test can't know which at authoring time), pair the factories that drive each
+version's arm so the test passes regardless of the toolchain — see
+`ForceLegacyCudaRegister` + `ForceLegacyIpcCapable` in `p2p-test.cc`.
 
 ## Dealing with each kind of dependency
